+++ Coosu.Storyboard.OsbX/Actions/ZoomOut.cs
@@ -8,7 +8,8 @@ namespace Coosu.Storyboard.OsbX.Actions
         {
         }
 
-        public ZoomOut(EasingType easing, float startTime, float endTime, float[] start, float[] end) : base(easing, startTime, endTime, start, end)
+        public ZoomOut(EasingType easing, double startTime, double endTime, double[] start, double[] end) 
+            : base(easing, startTime, endTime, start, end)
         {
         }
