+++ Coosu.Storyboard/Common/IEventHost.cs
@@ -4,10 +4,10 @@ namespace Coosu.Storyboard.Common
 {
     public interface IEventHost : IScriptable
     {
-        float MaxTime { get; }
-        float MinTime { get; }
-        float MaxStartTime { get; }
-        float MinEndTime { get; }
+        double MaxTime { get; }
+        double MinTime { get; }
+        double MaxStartTime { get; }
+        double MinEndTime { get; }
 
         bool EnableGroupedSerialization { get; set; }
         SortedSet<ICommonEvent> Events { get; }
