+++ Coosu.Storyboard/Extensibility/EventCreationDelegate.cs
@@ -3,6 +3,6 @@
 namespace Coosu.Storyboard.Extensibility
 {
     public delegate CommonEvent EventCreationDelegate(EventType e, EasingType easing,
-        float startTime, float endTime,
-        float[] start, float[] end);
+        double startTime, double endTime,
+        double[] start, double[] end);
 }
